requirements.txt
appdirs==1.4.4
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
decorator==5.1.1
fonttools==4.32.0
idna==3.3
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mne==1.0.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pooch==1.6.0
pyparsing==3.0.8
pyserial==3.5
python-dateutil==2.8.2
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tqdm==4.64.0
urllib3==1.26.9
